requirements.txt
# Gmail API
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# LLM APIs
anthropic>=0.39.0
openai>=1.54.0

# Environment management
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-dateutil>=2.8.2

# Optional: OCR for scanned PDFs
# pytesseract>=0.3.10
# Pillow>=10.0.0
